Matrix Recovery
See issue trezor/trezor-mcu#96

This adds a new type in the message RecoveryDevice, which can
be used by the host to signal which recovery procedures it supports.
It is a bitfield containing a bit for every supported procedure.

Keepkey already uses field 7 for their recovery type so we skip
this field id for easier compatibility.